fix all relevant warnings on MSVC x64

+++ cpp/lib/include/opendnp3/StackStatistics.h
@@ -33,16 +33,16 @@ struct StackStatistics
     struct Link
     {
         /// number of unexpected frames
-        uint32_t numUnexpectedFrame = 0;
+        uint64_t numUnexpectedFrame = 0;
 
         /// frames received w/ wrong master bit
-        uint32_t numBadMasterBit = 0;
+        uint64_t numBadMasterBit = 0;
 
         /// frames received for an unknown destination
-        uint32_t numUnknownDestination = 0;
+        uint64_t numUnknownDestination = 0;
 
         /// frames received for an unknown source
-        uint32_t numUnknownSource = 0;
+        uint64_t numUnknownSource = 0;
     };
 
     struct Transport
@@ -50,25 +50,25 @@ struct StackStatistics
         struct Rx
         {
             /// Number of valid TPDU's received
-            uint32_t numTransportRx = 0;
+            uint64_t numTransportRx = 0;
 
             /// Number of TPDUs dropped due to malformed contents
-            uint32_t numTransportErrorRx = 0;
+            uint64_t numTransportErrorRx = 0;
 
             /// Number of times received data was too big for reassembly buffer
-            uint32_t numTransportBufferOverflow = 0;
+            uint64_t numTransportBufferOverflow = 0;
 
             /// number of times transport buffer is discard due to new FIR
-            uint32_t numTransportDiscard = 0;
+            uint64_t numTransportDiscard = 0;
 
             /// number of segments ignored due to bad FIR/FIN or SEQ
-            uint32_t numTransportIgnore = 0;
+            uint64_t numTransportIgnore = 0;
         };
 
         struct Tx
         {
             /// Number of TPDUs transmitted
-            uint32_t numTransportTx = 0;
+            uint64_t numTransportTx = 0;
         };
 
         Transport() = default;
